Refactor drive types into multiple files.
These classes are due to get much longer when we add functionality to
them.
