+++ pepper2/daemon/controller.py
@@ -7,7 +7,8 @@ from gi.repository import GLib
 from pkg_resources import resource_string
 
 from pepper2 import __version__
-from pepper2.drives import DRIVE_TYPES, DriveGroup
+from pepper2.drive_types import DRIVE_TYPES
+from pepper2.drives import DriveGroup
 from pepper2.status import DaemonStatus
 
 LOGGER = logging.getLogger(__name__)
