+++ pepper2/daemon/udisks_manager.py
@@ -10,7 +10,9 @@ from typing import Dict, List, Type
 
 from pydbus.bus import Bus
 
-from pepper2.drives import DRIVE_TYPES, Constraint, Drive, DriveType
+from pepper2.constraint import Constraint
+from pepper2.drive_types import DRIVE_TYPES
+from pepper2.drives import Drive, DriveType
 
 from .controller import Controller
